# comandas-backend/openapi.yaml
openapi: 3.0.3
info:
  title: FoodFlow API
  description: |
    API REST del sistema de comandas FoodFlow.
    Sistema SaaS de comandas para locales gastronómicos pequeños.
    Arquitectura multi-tenant por fila — cada request requiere el header X-Local-Id.
  version: 1.0.0
  contact:
    name: Agustín Palma
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
security:
  - localId: []
...
paths:
...
  /api/mesas:
    get:
      tags: [Mesas]
      summary: Listar todas las mesas del local
      operationId: listarMesas
      responses:
        '200':
          description: Lista de mesas del local
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaResponse'
    post:
      tags: [Mesas]
      summary: Crear una nueva mesa
      operationId: crearMesa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearMesaRequest'
      responses:
        '201':
          description: Mesa creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesaResponse'
  /api/mesas/{mesaId}/abrir:
    post:
      tags: [Mesas]
      summary: Abrir una mesa libre y crear su pedido inicial
      operationId: abrirMesa
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      responses:
        '201':
          description: Mesa abierta y pedido creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbrirMesaResponse'
  /api/mesas/{mesaId}:
    delete:
      tags: [Mesas]
      summary: Eliminar una mesa (solo si está LIBRE)
      operationId: eliminarMesa
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      responses:
        '204':
          description: Mesa eliminada exitosamente
  /api/mesas/{mesaId}/pedido-actual:
    get:
      tags: [Mesas]
      summary: Consultar el detalle del pedido activo de una mesa
      description: "HU-06: Ver pedido de una mesa. Retorna ítems, totales y contexto."
      operationId: consultarPedidoActual
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      responses:
        '200':
          description: Detalle del pedido activo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetallePedidoResponse'
  /api/mesas/{mesaId}/cierre:
    post:
      tags: [Mesas]
      summary: Cerrar una mesa y finalizar su pedido activo
      description: |
        Evento de negocio crítico que consolida:
        - Cierre del pedido (inmutabilidad financiera)
        - Registro de pagos (soporte pagos parciales/split)
        - Liberación de la mesa (recurso físico)
      operationId: cerrarMesa
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarMesaRequest'
      responses:
        '200':
          description: Mesa cerrada con snapshot contable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CerrarMesaResponse'
  /api/mesas/{mesaId}/ticket:
    get:
      tags: [Mesas]
      summary: Obtener datos para impresión de ticket de venta (cliente)
      description: |
        HU-29: Ticket de venta para impresora térmica.
        Endpoint de solo lectura — no cierra mesa, no modifica estado.
        Contiene resumen financiero con totales, descuentos y datos del local.
      operationId: obtenerTicket
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      responses:
        '200':
          description: Datos estructurados del ticket para renderizado en frontend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketImpresionResponse'
  /api/mesas/{mesaId}/comanda:
    get:
      tags: [Mesas]
      summary: Obtener datos para impresión de comanda operativa (cocina/barra)
      description: |
        HU-05: Comanda operativa para cocina.
        Endpoint de solo lectura — no modifica pedido ni estado de mesa.
        No incluye valores monetarios.
      operationId: obtenerComanda
      parameters:
        - $ref: '#/components/parameters/mesaIdPath'
      responses:
        '200':
          description: Datos estructurados de la comanda para renderizado en frontend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComandaImpresionResponse'

  # ============================================================
  # PEDIDOS
  # ============================================================
...
components:
  securitySchemes:
    localId:
      type: apiKey
      in: header
      name: X-Local-Id
      description: UUID del local (tenant). Obligatorio en todos los endpoints.
  parameters:
    mesaIdPath:
      name: mesaId
      in: path
      required: true
      description: UUID de la mesa
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CrearMesaRequest:
      type: object
      required: [numero]
      properties:
        numero:
          type: integer
          description: Número de la nueva mesa
          example: 15
    MesaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la mesa
        numero:
          type: integer
          description: Número de la mesa
          example: 5
        estado:
          type: string
          enum: [LIBRE, ABIERTA]
          description: Estado actual de la mesa
    AbrirMesaResponse:
      type: object
      properties:
        mesaId:
          type: string
          format: uuid
        numeroMesa:
          type: integer
        estadoMesa:
          type: string
          enum: [LIBRE, ABIERTA]
        pedidoId:
          type: string
          format: uuid
        numeroPedido:
          type: integer
        estadoPedido:
          type: string
          enum: [ABIERTO, CERRADO]
        fechaApertura:
          type: string
          format: date-time
    CerrarMesaRequest:
      type: object
      properties:
        pagos:
          type: array
          description: Lista de pagos para cubrir el total del pedido
          items:
            $ref: '#/components/schemas/PagoRequest'
    PagoRequest:
      type: object
      properties:
        medio:
          type: string
          enum: [EFECTIVO, TARJETA, TRANSFERENCIA, QR, A_CUENTA]
          description: Medio de pago utilizado
        monto:
          type: number
          format: double
          description: Monto del pago (debe ser > 0)
          example: 5000.00
    CerrarMesaResponse:
      type: object
      properties:
        mesaId:
          type: string
          format: uuid
        mesaNumero:
          type: integer
        mesaEstado:
          type: string
          enum: [LIBRE, ABIERTA]
        pedidoId:
          type: string
          format: uuid
        pedidoEstado:
          type: string
          enum: [ABIERTO, CERRADO]
        montoSubtotal:
          type: number
          format: double
          description: Subtotal congelado (antes de descuentos)
        montoDescuentos:
          type: number
          format: double
          description: Monto total de descuentos aplicados
        montoTotal:
          type: number
          format: double
          description: Total final congelado
        pagos:
          type: array
          items:
            $ref: '#/components/schemas/PagoResponse'
        fechaCierre:
          type: string
          format: date-time
    PagoResponse:
      type: object
      properties:
        medio:
          type: string
          enum: [EFECTIVO, TARJETA, TRANSFERENCIA, QR, A_CUENTA]
        monto:
          type: number
          format: double
        fecha:
          type: string
          format: date-time
    DetallePedidoResponse:
      type: object
      properties:
        pedidoId:
          type: string
          format: uuid
        numeroPedido:
          type: integer
        numeroMesa:
          type: integer
        estado:
          type: string
          enum: [ABIERTO, CERRADO]
        fechaApertura:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemDetalleDTO'
        subtotal:
          type: number
          format: double
          description: Total sin descuentos
        totalDescuentos:
          type: number
          format: double
          description: Suma de descuentos de todos los ítems
        totalParcial:
          type: number
          format: double
          description: Lo que paga el cliente (subtotal - descuentos)
    ItemDetalleDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombreProducto:
          type: string
        cantidad:
          type: integer
        precioUnitarioBase:
          type: number
          format: double
          description: Precio de lista (para tachar en UI)
        subtotal:
          type: number
          format: double
          description: precioBase × cantidad
        descuentoTotal:
          type: number
          format: double
          description: Monto de descuento aplicado
        precioFinal:
          type: number
          format: double
          description: Lo que paga el cliente
        observacion:
          type: string
          nullable: true
        nombrePromocion:
          type: string
          nullable: true
          description: Nombre de la promo aplicada
        tienePromocion:
          type: boolean

    # ----------------------------------------------------------
    # PEDIDOS — Request/Response
    # ----------------------------------------------------------
...
    TicketImpresionResponse:
      type: object
      description: "HU-29: Datos estructurados para impresión de ticket de venta (cliente)."
      required: [header, items, totales, footer]
      properties:
        header:
          $ref: '#/components/schemas/HeaderTicket'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemTicket'
        totales:
          $ref: '#/components/schemas/TotalesTicket'
        footer:
          $ref: '#/components/schemas/FooterTicket'
    HeaderTicket:
      type: object
      required: [nombreLocal, fechaHora, numeroMesa]
      properties:
        nombreLocal:
          type: string
        direccion:
          type: string
          nullable: true
        telefono:
          type: string
          nullable: true
        cuit:
          type: string
          nullable: true
        fechaHora:
          type: string
          format: date-time
        numeroMesa:
          type: integer
    ItemTicket:
      type: object
      required: [cantidad, descripcion, precioUnitario, importe]
      properties:
        cantidad:
          type: integer
        descripcion:
          type: string
        precioUnitario:
          type: number
          format: double
        importe:
          type: number
          format: double
    TotalesTicket:
      type: object
      required: [subtotal, montoDescuentoPromos, montoDescuentoManual, totalFinal, desglosePromociones]
      properties:
        subtotal:
          type: number
          format: double
        montoDescuentoPromos:
          type: number
          format: double
        montoDescuentoManual:
          type: number
          format: double
        totalFinal:
          type: number
          format: double
        desglosePromociones:
          type: array
          items:
            $ref: '#/components/schemas/DesglosePromocion'
    DesglosePromocion:
      type: object
      required: [nombrePromocion, ahorro]
      properties:
        nombrePromocion:
          type: string
        ahorro:
          type: number
          format: double
    FooterTicket:
      type: object
      properties:
        mensajeBienvenida:
          type: string
          nullable: true
    ComandaImpresionResponse:
      type: object
      description: "HU-05: Datos estructurados para impresión de comanda operativa (cocina/barra). Sin valores monetarios."
      required: [header, items]
      properties:
        header:
          $ref: '#/components/schemas/HeaderComanda'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemComanda'
    HeaderComanda:
      type: object
      required: [numeroMesa, numeroPedido, fechaHora]
      properties:
        numeroMesa:
          type: integer
        numeroPedido:
          type: integer
        fechaHora:
          type: string
          format: date-time
    ItemComanda:
      type: object
      required: [cantidad, nombreProducto]
      properties:
        cantidad:
          type: integer
        nombreProducto:
          type: string
        observaciones:
          type: string
          nullable: true